+++ src/makeandship/elasticsearch/queries/query-builder.php
@@ -12,8 +12,7 @@ class QueryBuilder
     public function __construct()
     {
         // default include all valid post types
-        $post_mapping_builder = new PostMappingBuilder();
-        $this->post_types = $post_mapping_builder->get_valid_post_types();
+        $this->post_types = PostMappingBuilder::get_valid_post_types();
 
         $this->freetext = null;
         $this->fuzziness = null;
